refactor(#26): Repository 오타 수정
